deser integration tests
