+++ crates/rules/src/read.rs
@@ -102,53 +102,3 @@ fn read_rules_db(xs: Vec<RuleSource>) -> Result<DB, Error> {
 
     Ok(lint_db(DB::from_sources(lookup)))
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::read::deserialize_rules_db;
-    use std::error::Error;
-
-    #[test]
-    fn deser_simple() -> Result<(), Box<dyn Error>> {
-        let db = deserialize_rules_db(
-            r#"
-        [/foo.bar]
-        allow perm=any all : all
-        "#,
-        )?;
-        assert_eq!(db.len(), 1);
-        assert_eq!(db.rules().len(), 1);
-        assert!(db.rules().first().unwrap().valid);
-        Ok(())
-    }
-
-    #[test]
-    fn deser_set() -> Result<(), Box<dyn Error>> {
-        let db = deserialize_rules_db(
-            r#"
-        [/foo.bar]
-        %foo=bar
-        allow perm=any all : all
-        "#,
-        )?;
-        assert_eq!(db.len(), 2);
-        assert_eq!(db.rules().len(), 1);
-        assert_eq!(db.sets().len(), 1);
-        Ok(())
-    }
-
-    #[test]
-    fn deser_two_files() -> Result<(), Box<dyn Error>> {
-        let db = deserialize_rules_db(
-            r#"
-        [/foo.rules]
-        allow perm=execute all : all
-        [/bar.rules]
-        allow perm=any all : all
-        "#,
-        )?;
-        assert_eq!(db.len(), 2);
-        assert_eq!(db.rules().len(), 2);
-        Ok(())
-    }
-}
